A better patch for the problem described at https://s.afterlogic.com/forum/forum_posts.asp?TID=7995&PN=1&TPN=1
This updated patch addresses the same problem, but does so more elegantly by moving the existing code that handles buffering of data larger than $iCount so that it applies to the "if (\is_string($mCurrentPart))" case too.

This will make the code correctly handle a case where $mCurrentPart is > 8192 bytes to start with, which neither the original code nor the first version of the patch handled.